Funny Story Generator
I accidentally committed and uncommitted a bunch of times.  FINALLY:
This is a simple example of C++ code in which a user enters a few names
and then the code can generate a funny story based on a randomly
generated number.  It's super simply but it was fun to make.