Added a GUESSING game
I created a game that lets you guess a number in-between 1 and 10.  If
the number is too high the program tells you that your guess is too
high.  Similarly if the number you guess is too low the program tells
you that your guess is too low.  Once your guess is right you can opt to
play again or close the program.