Added Inches to Centimeters Converter
This is a simple program that has a method that converts inches to
centimeters.
Each inch is 2.54 centimeters so the program takes a user input of
inches and multiplies it by 2.54 and then returns the value.
Additionally I put in some simple error handling in case the user input
was not a number.
The WHILE loops ensure that the user will insert correct data and that
the program can loop as many times as the user likes.